+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Changed
+- Change getNativeDateTimeFormat to handle column schema to detect detailed datetime format
 
 ## [0.7.0] - 2017-09-15
 ### Added